feat(ui): separate Add KVM form into LXD and virsh forms
