+++ ui/src/app/kvm/views/KVMList/AddKVM/AddKVM.test.tsx
@@ -1,0 +1,100 @@
+import { mount } from "enzyme";
+import { Provider } from "react-redux";
+import { MemoryRouter } from "react-router-dom";
+import configureStore from "redux-mock-store";
+
+import AddKVM from "./AddKVM";
+
+import { PodType } from "app/store/pod/types";
+import type { RootState } from "app/store/root/types";
+import {
+  configState as configStateFactory,
+  generalState as generalStateFactory,
+  powerField as powerFieldFactory,
+  powerTypesState as powerTypesStateFactory,
+  powerType as powerTypeFactory,
+  podState as podStateFactory,
+  resourcePool as resourcePoolFactory,
+  resourcePoolState as resourcePoolStateFactory,
+  rootState as rootStateFactory,
+  zone as zoneFactory,
+  zoneState as zoneStateFactory,
+} from "testing/factories";
+
+const mockStore = configureStore();
+
+describe("AddKVM", () => {
+  let initialState: RootState;
+
+  beforeEach(() => {
+    initialState = rootStateFactory({
+      config: configStateFactory({
+        items: [{ name: "maas_name", value: "MAAS" }],
+      }),
+      general: generalStateFactory({
+        powerTypes: powerTypesStateFactory({
+          data: [
+            powerTypeFactory({
+              name: PodType.VIRSH,
+              fields: [
+                powerFieldFactory({ name: "power_address" }),
+                powerFieldFactory({ name: "power_pass" }),
+              ],
+            }),
+          ],
+          loaded: true,
+        }),
+      }),
+      pod: podStateFactory({
+        loaded: true,
+      }),
+      resourcepool: resourcePoolStateFactory({
+        items: [resourcePoolFactory()],
+        loaded: true,
+      }),
+      zone: zoneStateFactory({
+        items: [zoneFactory()],
+        loaded: true,
+      }),
+    });
+  });
+
+  it("fetches the necessary data on load", () => {
+    const state = { ...initialState };
+    const store = mockStore(state);
+    mount(
+      <Provider store={store}>
+        <MemoryRouter
+          initialEntries={[{ pathname: "/kvm/add", key: "testKey" }]}
+        >
+          <AddKVM />
+        </MemoryRouter>
+      </Provider>
+    );
+    const expectedActions = [
+      "FETCH_GENERAL_POWER_TYPES",
+      "resourcepool/fetch",
+      "zone/fetch",
+    ];
+    const actions = store.getActions();
+    expectedActions.forEach((expectedAction) => {
+      expect(actions.some((action) => action.type === expectedAction));
+    });
+  });
+
+  it("displays a spinner if data has not loaded yet", () => {
+    const state = { ...initialState };
+    state.resourcepool.loaded = false;
+    const store = mockStore(state);
+    const wrapper = mount(
+      <Provider store={store}>
+        <MemoryRouter
+          initialEntries={[{ pathname: "/kvm/add", key: "testKey" }]}
+        >
+          <AddKVM />
+        </MemoryRouter>
+      </Provider>
+    );
+    expect(wrapper.find("Spinner").length).toBe(1);
+  });
+});
