+++ ui/src/app/kvm/views/KVMList/AddKVM/AddKVM.tsx
@@ -1,0 +1,48 @@
+import { useEffect, useState } from "react";
+
+import { Spinner } from "@canonical/react-components";
+import { useDispatch, useSelector } from "react-redux";
+
+import AddLxd from "./AddLxd";
+import AddVirsh from "./AddVirsh";
+
+import { general as generalActions } from "app/base/actions";
+import { useWindowTitle } from "app/base/hooks";
+import generalSelectors from "app/store/general/selectors";
+import { PodType } from "app/store/pod/types";
+import { actions as resourcePoolActions } from "app/store/resourcepool";
+import resourcePoolSelectors from "app/store/resourcepool/selectors";
+import { actions as zoneActions } from "app/store/zone";
+import zoneSelectors from "app/store/zone/selectors";
+
+export type SetKvmType = (kvmType: PodType) => void;
+
+export const AddKvmTypeSelect = (): JSX.Element => {
+  const dispatch = useDispatch();
+  const powerTypesLoaded = useSelector(generalSelectors.powerTypes.loaded);
+  const resourcePoolsLoaded = useSelector(resourcePoolSelectors.loaded);
+  const zonesLoaded = useSelector(zoneSelectors.loaded);
+  const allLoaded = powerTypesLoaded && resourcePoolsLoaded && zonesLoaded;
+
+  const [kvmType, setKvmType] = useState<PodType>(PodType.VIRSH);
+
+  useEffect(() => {
+    dispatch(generalActions.fetchPowerTypes());
+    dispatch(resourcePoolActions.fetch());
+    dispatch(zoneActions.fetch());
+  }, [dispatch]);
+
+  useWindowTitle("Add KVM");
+
+  if (!allLoaded) {
+    return <Spinner className="u-no-margin u-no-padding" text="Loading" />;
+  }
+  return (
+    <>
+      {kvmType === PodType.LXD && <AddLxd setKvmType={setKvmType} />}
+      {kvmType === PodType.VIRSH && <AddVirsh setKvmType={setKvmType} />}
+    </>
+  );
+};
+
+export default AddKvmTypeSelect;
