+++ ui/src/app/kvm/views/KVMList/AddKVM/AddLxd/AddLxdFields/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./AddLxdFields";
