+++ ui/src/app/kvm/views/KVMList/AddKVM/AddVirsh/AddVirshFields/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./AddVirshFields";
