+++ ui/src/app/kvm/views/KVMList/AddKVM/AddVirsh/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./AddVirsh";
