+++ ui/src/app/kvm/views/KVMList/AddKVM/KvmTypeSelect/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./KvmTypeSelect";
