+++ ui/src/app/kvm/views/KVMList/AddKVM/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./AddKVM";
+export type { SetKvmType } from "./AddKVM";
