--- ui/src/app/kvm/views/KVMList/AddKVMForm/AddKVMFormFields/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AddKVMFormFields";
