--- ui/src/app/kvm/views/KVMList/AddKVMForm/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AddKVMForm";
